requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mysql-connector-python==8.3.0
pydantic==2.5.3
python-dotenv==1.0.1
